% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoedit-package.R
\docType{package}
\name{autoedit-package}
\alias{autoedit}
\alias{autoedit-package}
\title{autoedit: Collaborative Code Editor Widget for Automerge}
\description{
A real-time collaborative code editor widget built on CodeMirror 6 and Automerge CRDT. Connects via WebSocket to any automerge-repo compatible sync server for seamless multi-user editing with automatic conflict resolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shikokuchuo/autoedit}
  \item \url{https://shikokuchuo.net/autoedit/}
  \item \url{http://shikokuchuo.net/autoedit/}
  \item Report bugs at \url{https://github.com/shikokuchuo/autoedit/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
