% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editor-shiny}
\alias{editor-shiny}
\alias{editor_output}
\alias{render_editor}
\title{Shiny bindings for editor}
\usage{
editor_output(outputId, width = "100\%", height = "400px")

render_editor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Widget dimensions (must be valid CSS unit or a number
which will be coerced to a string and have "px" appended).}

\item{expr}{An expression that generates an editor widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Logical, whether \code{expr} is a quoted expression.}
}
\value{
\code{editor_output()} returns a Shiny output element.
\code{render_editor()} returns a Shiny render function.
}
\description{
Output and render functions for using editor within Shiny
applications and interactive R Markdown documents.
}
